feat: implement navigation guard for unsaved changes handling
